first

+++ src/main/kotlin/lesson1/task4.kt
@@ -4,8 +4,8 @@ fun main() {
     val distance: Long = 40868600000
     val age: Byte = 27
     val partOfTheDay: Float = 0.075F
-    val second: Int = 6480
-    val partOfTheYear:Double = 2.0547945205479453E-4
+    val second: Short = 6480
+    val partOfTheYear: Float = 2.0547945205479453E-4F
     val apogeeOfTheOrbit: Int = 327000
 
     println(distance)
